authorization + showing userEmail

+++ backend/routes/posts.js
@@ -40,7 +40,8 @@ router.post(
   const post = new Post({
     title: req.body.title,
     content: req.body.content,
-    imagePath: url + "/images/" + req.file.filename
+    imagePath: url + "/images/" + req.file.filename,
+    creator: req.userData.userId
   });
   post.save()   // this is mongoose
     .then(createdPost => { 
@@ -77,12 +78,17 @@ router.put(
     _id: req.body.id,
     title: req.body.title,
     content: req.body.content,
-    imagePath: imagePath
+    imagePath: imagePath,
+    creator: req.userData.userId
   });
   console.log(post);
-  Post.updateOne({ _id: req.params.id }, post).then(result => {
+  Post.updateOne({ _id: req.params.id, creator: req.userData.userId }, post).then(result => {
     // console.log(result);
-    res.status(200).json({ message: "Update successful!" });
+    if (result.nModified > 0) {
+      res.status(200).json({ message: "Update successful!" });
+    } else {
+      res.status(401).json({ message: "Not authorized! Update not successful!" });
+    }
   });
 });
 
@@ -132,10 +138,14 @@ router.get("/:id", (req, res, next) => {
 });
 
 router.delete("/:id", checkAuth, (req, res, next) => {
-  Post.deleteOne({ _id: req.params.id })    // this is mongoose
+  Post.deleteOne({ _id: req.params.id, creator: req.userData.userId })    // this is mongoose
   .then(result => {
     console.log(result);
-    res.status(200).json({ message: "Post deleted!" });
+    if (result.n > 0) {
+      res.status(200).json({ message: "Deletion successful!" });
+    } else {
+      res.status(401).json({ message: "Not authorized! Deletion not successful!" });
+    }
   });
 });
